requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
paho-mqtt==2.1.0
pymongo==4.8.0
python-dotenv==1.0.1
